change price type to a decimal with scale of 2

+++ api/src/db/seeder/factories/product.factory.ts
@@ -7,7 +7,7 @@ export class ProductFactory extends Factory<Product> {
   protected definition(faker: Faker): Partial<Product> {
     return {
       name: faker.commerce.productName(),
-      price: Number.parseInt(faker.commerce.price(10, 100)),
+      price: Number.parseFloat(faker.commerce.price(10, 100)),
       image: faker.image.fashion(640, 480, true),
     };
   }
